refactor(components): decompose UserList into modular components
Added:

- Create AutoUserCreation.vue and RandomUserDisplay.vue
- Extract userDataGenerator utility
- Improve component separation and reusability

# 8
